- breaking down focus logic into easily readable segments - removed a small amount of duplication on Screen capture code - added previous killers as part of the priority focus list - added end state detection to work from Match awards screen because the top timer is not visible after the game has ended - State loop now checks for a timer, if the timer is near the end gate time, it will check for the match awards screen - Added a check for S3 download so the same replay is not downloaded twice if already in temp cache - Updated the README

+++ src/HeroesReplay.Replays/StormReplayHotsApiProvider.cs
@@ -59,14 +59,17 @@ namespace HeroesReplay.Replays
 
                             string stormReplay = Path.Combine(Path.GetTempPath(), request.Key);
 
-                            using (GetObjectResponse response = await s3Client.GetObjectAsync(request, provider.Token))
+                            if (!File.Exists(stormReplay))
                             {
-                                await using (MemoryStream memoryStream = new MemoryStream()) 
+                                using (GetObjectResponse response = await s3Client.GetObjectAsync(request, provider.Token))
                                 {
-                                    await response.ResponseStream.CopyToAsync(memoryStream);
-                                    logger.LogInformation($"[HOTSAPI][SAVING][{hotsApiReplay.Id}][{hotsApiReplay.Url}][{stormReplay}]");
-                                    await File.WriteAllBytesAsync(stormReplay, memoryStream.ToArray());
-                                    logger.LogInformation($"[HOTSAPI][SAVED][{hotsApiReplay.Id}][{hotsApiReplay.Url}][{stormReplay}]");
+                                    await using (MemoryStream memoryStream = new MemoryStream())
+                                    {
+                                        await response.ResponseStream.CopyToAsync(memoryStream);
+                                        logger.LogInformation($"[HOTSAPI][SAVING][{hotsApiReplay.Id}][{hotsApiReplay.Url}][{stormReplay}]");
+                                        await File.WriteAllBytesAsync(stormReplay, memoryStream.ToArray());
+                                        logger.LogInformation($"[HOTSAPI][SAVED][{hotsApiReplay.Id}][{hotsApiReplay.Url}][{stormReplay}]");
+                                    }
                                 }
                             }
 
@@ -118,7 +121,7 @@ namespace HeroesReplay.Replays
                     {
                         // Why?
                     }
-                    
+
 
                     logger.LogInformation($"[REPLAY][ERROR][Could not find suitable replay]");
                 }
